Split up editor module
