+++ lapce-ui/src/editor/gutter.rs
@@ -1,0 +1,97 @@
+use druid::{
+    BoxConstraints, Env, Event, EventCtx, LayoutCtx, LifeCycle, LifeCycleCtx,
+    PaintCtx, Size, UpdateCtx, Widget, WidgetId,
+};
+use lapce_data::data::LapceTabData;
+
+pub struct LapceEditorGutter {
+    view_id: WidgetId,
+    width: f64,
+}
+
+impl LapceEditorGutter {
+    pub fn new(view_id: WidgetId) -> Self {
+        Self {
+            view_id,
+            width: 0.0,
+        }
+    }
+}
+
+impl Widget<LapceTabData> for LapceEditorGutter {
+    fn event(
+        &mut self,
+        _ctx: &mut EventCtx,
+        _event: &Event,
+        _data: &mut LapceTabData,
+        _env: &Env,
+    ) {
+    }
+
+    fn lifecycle(
+        &mut self,
+        _ctx: &mut LifeCycleCtx,
+        _event: &LifeCycle,
+        _data: &LapceTabData,
+        _env: &Env,
+    ) {
+    }
+
+    fn update(
+        &mut self,
+        _ctx: &mut UpdateCtx,
+        _old_data: &LapceTabData,
+        _data: &LapceTabData,
+        _env: &Env,
+    ) {
+        // let old_last_line = old_data.buffer.last_line() + 1;
+        // let last_line = data.buffer.last_line() + 1;
+        // if old_last_line.to_string().len() != last_line.to_string().len() {
+        //     ctx.request_layout();
+        //     return;
+        // }
+
+        // if (*old_data.main_split.active == self.view_id
+        //     && *data.main_split.active != self.view_id)
+        //     || (*old_data.main_split.active != self.view_id
+        //         && *data.main_split.active == self.view_id)
+        // {
+        //     ctx.request_paint();
+        // }
+
+        // if old_data.editor.cursor.current_line(&old_data.buffer)
+        //     != data.editor.cursor.current_line(&data.buffer)
+        // {
+        //     ctx.request_paint();
+        // }
+
+        // if old_data.current_code_actions().is_some()
+        //     != data.current_code_actions().is_some()
+        // {
+        //     ctx.request_paint();
+        // }
+    }
+
+    fn layout(
+        &mut self,
+        ctx: &mut LayoutCtx,
+        bc: &BoxConstraints,
+        data: &LapceTabData,
+        _env: &Env,
+    ) -> Size {
+        let data = data.editor_view_content(self.view_id);
+        let last_line = data.buffer.last_line() + 1;
+        let char_width = data.config.editor_text_width(ctx.text(), "W");
+        self.width = (char_width * last_line.to_string().len() as f64).ceil();
+        let mut width = self.width + 16.0 + char_width * 2.0;
+        if data.editor.compare.is_some() {
+            width += self.width + char_width * 2.0;
+        }
+        Size::new(width.ceil(), bc.max().height)
+    }
+
+    fn paint(&mut self, ctx: &mut PaintCtx, data: &LapceTabData, _env: &Env) {
+        let data = data.editor_view_content(self.view_id);
+        data.paint_gutter(ctx, self.width);
+    }
+}
