+++ lapce-ui/src/find.rs
@@ -9,7 +9,7 @@ use lapce_data::{
     data::LapceTabData,
 };
 
-use crate::{editor::LapceEditorView, svg::get_svg, tab::LapceIcon};
+use crate::{editor::view::LapceEditorView, svg::get_svg, tab::LapceIcon};
 
 pub struct FindBox {
     input_width: f64,
