+++ lapce-ui/src/source_control.rs
@@ -19,7 +19,7 @@ use lapce_data::{
 use lapce_rpc::source_control::FileDiff;
 
 use crate::{
-    editor::LapceEditorView,
+    editor::view::LapceEditorView,
     panel::{LapcePanel, PanelHeaderKind},
     svg::{file_svg_new, get_svg},
 };
